requirements.txt
python-dotenv
streamlit
openai
faiss-cpu
PyPDF2
langchain
langchain-community
tiktoken